refactor: model -> models, controller -> controllers
