+++ BackEnd/src/models/meeting/meetingDetail.js
